refactor(views): 抽 showMultiPlatformSyncModal 到 views/publish-modal/multi-platform.js
Phase 3 第三步。把 AppleStyleView.showMultiPlatformSyncModal (~318 行) 完整
抽到独立模块，导出 showMultiPlatformPublishModal(view, options)。

input.js 里只剩一行薄包装（保留方法名以向后兼容外部调用方）：

    async showMultiPlatformSyncModal(options = {}) {
      return showMultiPlatformPublishModal(this, options);
    }

注意点：
- 新模块对 obsidian 用 const obsidian = require('obsidian') 而非
  destructure，因为 tests/multi_platform_modal.test.js 用
  obsidian.Modal = CapturingModal 在模块加载之后做 monkey-patch；
  destructure 会捕获原始 Modal 引用导致测试失败。Modal 改成
  new obsidian.Modal(...) 在调用时才解析。
- isMobileClient 因为只用一次，直接 inline 在新模块里，避免再加共享 helper。
- view.* 方法（preparePublishModalShell / prepareHtmlForWechatsyncArticle /
  getPublishContextFile / showWechatsyncEnqueueAcceptedModal /
  showMultiPlatformSyncResultModal 等）保留在 AppleStyleView，新模块只承担
  UI 壳。

input.js 7068 → 6754 行 (-314)。51 测试文件 / 594 用例全过，含 7 个发布弹窗
DOM 契约测试。
